feature: filter slangs
