+++ src/app/(home)/(posts)/_components/forum/PostCard.tsx
@@ -3,7 +3,8 @@ import Dot from '@/assets/images/common/Dot';
 import Share from '@/assets/images/common/Share';
 import BookmarkButton from '@/components/common/BookmarkButton';
 import LikeButton from '@/components/common/LikeButton';
-import { cutText } from '@/components/common/MarkdownCut';
+import { cutText, filterSlang } from '@/utils/markdownCut';
+import TagBlock from '@/components/common/TagBlock';
 import { handleLinkCopy } from '@/components/handleLinkCopy';
 import { PostCardProps } from '@/types/posts/forumTypes';
 import { timeForToday } from '@/utils/timeForToday';
@@ -55,7 +56,7 @@ const PostCard = ({ post }: PostCardProps) => {
             <Image src={post.thumbnail} alt="Post Thumbnail" width={300} height={300} objectFit="cover" />
           )}
         </div>
-        <h2 className="text-h4 font-bold text-neutral-900 mt-3">{post.title}</h2>
+        <h2 className="text-h4 font-bold text-neutral-900 mt-3">{filterSlang(post.title)}</h2>
         <div className="post-content mt-2 custom-markdown" data-color-mode="light">
           <MDEditor.Markdown source={cutText(processedContent, 500)} />
         </div>
